# openapi.yaml
openapi: 3.0.2
info:
  title: 'Profiler Buddy'
  version: '1.0.0'
paths:
  /profile:
    post:
      operationId: profile
      description: Predicts demographic data based on the user feed.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comments'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemographicData'
          description: ''
      tags:
      - Profile
      
  /questionnaire:
    post:
      operationId: questionnaire
      description: Predicts the answers of Beck's depression inventory.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comments'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'
          description: 'Beck depression inventory answers and contexts'
      tags:
      - Questionnaire

  /qa:
    post:
      operationId: question_answering
      description: Predicts the answer of given question in the context.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comments'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'
          description: 'Beck depression inventory answers and contexts'
      tags:
      - QuestionAnswering

components:
  schemas:
    Comments:
      type: object
      properties:
        experiment_id:
          type: string
          example: 'profile32387'
        comments:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              text:
                type: string
                example: "I'm so tired of just feel empty, lonely, worthless, stupid, and a million other terrible things. But I'm not bad enough to end it yet. I just want it to stop so bad. I feel terrible all the time. I just want it to go away. Over the past month or so I've been thinking about death and wanting to die a lot more than normally. Like all I feel like listening to is songs about death or other sad things. I also only really like reading sad things now, or looking at sad things. And ive always like had thoughts about wanting to die but like not super seriously I guess. I also think about if people would care or not. And lately I've been like seriously thinking about how I would do it and where and stuff, which is something I've never done before in the years I've been thinking about killing myself. Recently I feel like I hate existing more and more every single day. I feel like im just waiting around for it to end. I don't care about having goals, or dreams. I just want to get through the shittiness of existence. I don't have motivation or really care about anything anymore. I just want it to be over."
            required:
            - date
            - text
      required:
        - comments
    DemographicData:
      type: object
      properties:
        gender:
          type: string
        age:
          type: string
    Questionnaire:
      type: object
      properties:
        questionnaire:
          type: object
          properties:
            q1:
              type: string
              example: '2'
            q2:
              type: string
              example: '0'
            q3:
              type: string
              example: '3'
            q18:
              type: string
              example: '1b'
            q19:
              type: string
              example: '2'
            q20:
              type: string
              example: '3'
            q21:
              type: string
              example: '1'
        questionnaire_reasons:
          type: object
          properties:
            q1:
              $ref: '#/components/schemas/QuestionnaireReason'
            q2:
              $ref: '#/components/schemas/QuestionnaireReason'
            q3:
              $ref: '#/components/schemas/QuestionnaireReason'
            q21:
              $ref: '#/components/schemas/QuestionnaireReason'
    QuestionnaireReason:
      type: object
      properties:
        context: 
          type: string
          example: 'I just want it to stop so bad. I feel terrible all the time.'
        score: 
          type: string
          example: '34872'
      